# specs/003-user-auth/contracts/auth-api.openapi.yaml
openapi: 3.0.3
info:
  title: Collab Todo - Authentication & Authorization API
  description: |
    API specification for user authentication and authorization features in the collaborative to-do application.
    
    ## Authentication
    - Uses JWT tokens stored in HTTP-only cookies
    - Session expiration: 24 hours
    - Cookie name: `auth_token`
    
    ## Authorization Hierarchy
    1. **Owner** (authenticated user who created list): Full control (read, write, delete)
    2. **Edit Link** (shareId): Read and write access (no delete)
    3. **View Link** (viewId): Read-only access
    
    ## Rate Limiting
    - Signup: 3 attempts per IP per hour
    - Signin: 5 attempts per email per 15 minutes
    
    ## Security
    - All endpoints use HTTPS in production
    - Passwords hashed with bcrypt (cost factor 12)
    - Sensitive endpoints require authentication
  version: 1.0.0
  contact:
    name: Collab Todo API Support
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.collabtodo.app
    description: Production server
...
paths:
...
  /api/lists:
    post:
      tags:
        - Lists
      summary: Create a new todo list (requires authentication)
      description: |
        **UPDATED**: Now requires authentication. Created list is associated with authenticated user as owner.
        
        **Authorization**: Must be authenticated
      operationId: createList
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 200
                  example: Team Sprint Tasks
      responses:
        '201':
          description: List created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  list:
                    $ref: '#/components/schemas/TodoList'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Authentication required

    get:
      tags:
        - Lists
      summary: Get all lists owned by authenticated user
      description: |
        **NEW**: Returns lists where authenticated user is the owner.
        
        Used for user dashboard view.
        
        **Authorization**: Must be authenticated
      operationId: getOwnedLists
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Lists retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  lists:
                    type: array
                    items:
                      $ref: '#/components/schemas/TodoList'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: JWT token stored in HTTP-only cookie
  schemas:
...
    TodoList:
      type: object
      required:
        - _id
        - shareId
        - viewId
        - title
        - createdAt
        - updatedAt
      properties:
        _id:
          type: string
          format: objectId
          example: 65a1b2c3d4e5f6a7b8c9d0e2
        userId:
          type: string
          format: objectId
          example: 65a1b2c3d4e5f6a7b8c9d0e1
          description: Owner user ID (optional for old lists created before auth)
        shareId:
          type: string
          pattern: '^[a-zA-Z0-9]{9}$'
          example: aBc123XyZ
          description: Edit link identifier (9-char alphanumeric)
        viewId:
          type: string
          pattern: '^[a-zA-Z0-9]{9}$'
          example: xYz987AbC
          description: View-only link identifier (9-char alphanumeric)
        title:
          type: string
          maxLength: 200
          example: Team Sprint Tasks
        createdAt:
          type: string
          format: date-time
          example: '2026-01-06T11:00:00.000Z'
        updatedAt:
          type: string
          format: date-time
          example: '2026-01-06T11:00:00.000Z'
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Invalid email format
          description: Human-readable error message